% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_amino_acid.R
\name{analyze_amino_acid}
\alias{analyze_amino_acid}
\title{Title Analyze amino acid in protein sequence}
\usage{
analyze_amino_acid(sequence, amino_acid)
}
\arguments{
\item{sequence}{A character string with the protein sequence}

\item{amino_acid}{A single letter amino acid}
}
\value{
A list with count and positions of the provided amino acid
}
\description{
Title Analyze amino acid in protein sequence
}
\examples{
analyze_amino_acid("MKTWLTKK", "K")

}
