% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_protein.R
\name{is_valid_protein}
\alias{is_valid_protein}
\title{Title Is rotein Sequence is valid?}
\usage{
is_valid_protein(sequence)
}
\arguments{
\item{sequence}{A character string representing a protein sequence}
}
\value{
A logical value: \code{TRUE} if all characters are valid amino acids,
\code{FALSE} otherwise.
}
\description{
Title Is rotein Sequence is valid?
}
\examples{
is_valid_protein("ACDEFGHIKLMNPQRSTVWY")  # TRUE
is_valid_protein("ACDEX")                 # FALSE

}
