% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_mass.R
\name{protein_mass}
\alias{protein_mass}
\title{Calculate mass of a protein}
\usage{
protein_mass(sequence)
}
\arguments{
\item{sequence}{A character string representing a protein sequence, using
one letter amino acids.}
}
\value{
A numeric value, the total molecular mass of the protein sequence
in Dalton (Da).
}
\description{
Calculate mass of a protein
}
\examples{
protein_mass("ACDE")
protein_mass("MALWMRLLPLLALLALWGPDPAAA")
}
