% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_dna.R
\name{translate_dna}
\alias{translate_dna}
\title{Title Translate DNA sequence into protein sequence}
\usage{
translate_dna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence}
}
\value{
A character string representing the translated protein sequence
}
\description{
Title Translate DNA sequence into protein sequence
}
\examples{
translate_dna("ATGGCCATTGTAATGGGCCGCTGAAAGGGTGCCCGATAG")

}
